+++ flask_unchained/_code_templates/project/app/graphql/schema.py
@@ -1,6 +1,6 @@
 import graphene
 
-from flask_unchained.bundles.graphene import MutationObjectType, QueryObjectType
+from flask_unchained.bundles.graphene import MutationsObjectType, QueriesObjectType
 
 from . import types
 from . import mutations