+++ tests/bundles/graphene/_bundles/graphene_bundle/graphql/schema.py
@@ -1,12 +1,12 @@
 import graphene
 
-from flask_unchained.bundles.graphene import MutationObjectType, QueryObjectType
+from flask_unchained.bundles.graphene import MutationsObjectType, QueriesObjectType
 
 from . import types
 from . import mutations
 
 
-class GrapheneBundleQueries(QueryObjectType):
+class GrapheneBundleQueries(QueriesObjectType):
     parent = graphene.Field(types.Parent, id=graphene.ID(required=True))
     parents = graphene.List(types.Parent)
 
@@ -14,7 +14,7 @@ class GrapheneBundleQueries(QueryObjectType):
     children = graphene.List(types.Child)
 
 
-class GrapheneBundleMutations(MutationObjectType):
+class GrapheneBundleMutations(MutationsObjectType):
     create_parent = mutations.CreateParent.Field()
     delete_parent = mutations.DeleteParent.Field()
     edit_parent = mutations.EditParent.Field()