+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - rename `SQLAlchemy` extension class to `SQLAlchemyUnchained`
 - remove `Bundle.root_folder` descriptor as it made no sense
 - rename `ConfigPropertyMeta` to `ConfigPropertyMetaclass`
+- renamed the Graphene Bundle's `QueryObjectType` to `QueriesObjectType` and `MutationObjectType` to `MutationsObjectType`
 
 ### Bug fixes